リファクタ: 評価層 counter の参照解決・整形を補助モジュールへ分離する
## 概要

`counter.rs` に同居していた pass2 の参照解決とテンプレート整形のフリー関数を
責務単位のサブモジュールへ切り出す。振る舞いは変えない。

## 変更内容

`crates/parser/src/evaluator/counter.rs`（本体約 515 行）を `mod.rs` を使わない
2018 スタイルで分割し、`counter.rs` + `counter/{resolve,format}.rs` 構成にした。

- `counter/resolve.rs` — pass2 の参照解決（`resolve_refs` / `resolve_list_item` /
  `resolve_proof_target` / `resolve_inlines`）。`CounterRegistry::resolve_label`
  のみを呼ぶ木探索で、private フィールドに非依存
- `counter/format.rs` — テンプレート整形の純粋関数（`parse_counter_name` /
  `expand_ref_format` / `expand_placeholders`）
- 親 `counter.rs` には `CounterRegistry` 構造体 + impl、`ResolvedLabel`、
  `theorem_reset_level` を残す

`resolve_refs` は `lib.rs` / `theorem.rs` から `counter::resolve_refs` として
外部参照されているため、親で `pub(crate) use resolve::resolve_refs;` と
再エクスポートして呼び出しパスを維持した。`format.rs` の関数は親からのみ
呼ばれるため `pub(super)`。

## テスト

- `cargo test -p parser` — counter 関連含め全 green
- `cargo test`（workspace） — 全 green
- `cargo clippy --all-targets` — warning なし
- `cargo +nightly fmt --check` — 差分なし

## スコープ外

`CounterRegistry` impl 自体の内部分割（private フィールドに密結合のため対象外）。
採番・参照書式の挙動変更。

## 関連

Closes #147
